requirements.txt
aiohttp==1.0.1
async-timeout==1.0.0
certifi==2016.8.31
chardet==2.3.0
future==0.15.2
multidict==2.1.0
pymongo==3.3.0
python-telegram-bot==5.0.0
pytz==2016.6.1
requests==2.11.1
urllib3==1.17
